Fixed bug in determining if PFS is on or not - was in runAcquisition - due to change in property value name.
Bug in visitZ also corrected - drift was added to set z position even if no point was input - i.e. set z was just the measured z position - so drift would already be included in that.

Minor changes to various scripts - mostly removing code that has been commented out

Added code to stop acquisition within the positions loop if the user clicks the stop button

Fixed a bug in correctDrift - was failing because of the change in property value from 'Locked in focus' to 'Locked'. Also now waits for focus before measuring drift.

Additions to the gui when choosing switching method. Now displays all of the pumping parameters after entry with a graph of flow rates.